+++ flask_logconfig/__init__.py
@@ -4,9 +4,11 @@
 import logging
 from collections import defaultdict
 import contextlib
+import datetime
 
 import logconfig
 
+import flask
 from flask import (
     current_app,
     request,
@@ -119,7 +121,8 @@ class LogConfig(object):
                              handler_class)
 
         if app.config['LOGCONFIG_REQUESTS_ENABLED']:
-            app.after_request(self.after_request_handler)
+            app.before_request(self.before_request)
+            app.after_request(self.after_request)
 
     def setup_logging(self, app):
         """Setup logging configuration for application."""
@@ -190,21 +193,35 @@ class LogConfig(object):
         for listener in self.get_listeners(app).values():
             listener.stop()
 
-    def after_request_handler(self, response):
+    def before_request(self):
+        """Store information related to start of request."""
+        flask.g.logconfig = {
+            'start': datetime.datetime.now()
+        }
+
+    def after_request(self, response):
         """Log request."""
+        logger = self.get_requests_logger()
+        data = self.get_request_message_data(response)
+        logger.log(self.config['LOGCONFIG_REQUESTS_LEVEL'],
+                   self.make_request_message(data),
+                   extra={'request': request,
+                          'response': response,
+                          'execution_time': data.get('execution_time')})
+
+        return response
+
+    def get_requests_logger(self):
+        """Get designated logger for requests."""
         if self.config['LOGCONFIG_REQUESTS_LOGGER']:
             logger = logging.getLogger(
                 self.config['LOGCONFIG_REQUESTS_LOGGER'])
         else:
             logger = self.get_app().logger
 
-        logger.log(self.config['LOGCONFIG_REQUESTS_LEVEL'],
-                   self.make_request_msg(response),
-                   extra={'response': response, 'request': request})
+        return logger
 
-        return response
-
-    def get_request_msg_data(self, response):
+    def get_request_message_data(self, response):
         """Return data for use in request message format string."""
         data = {}
 
@@ -224,7 +241,8 @@ class LogConfig(object):
         # Update with response data.
         data.update({
             'status_code': response.status_code,
-            'status': response.status
+            'status': response.status,
+            'execution_time': self.get_execution_time()
         })
 
         session_data = defaultdict(lambda: None)
@@ -237,10 +255,23 @@ class LogConfig(object):
 
         return data
 
-    def make_request_msg(self, response):
+    def make_request_message(self, data):
         """Return string formatted message for request log message."""
-        return (self.config['LOGCONFIG_REQUESTS_MSG_FORMAT']
-                .format(**self.get_request_msg_data(response)))
+        return self.config['LOGCONFIG_REQUESTS_MSG_FORMAT'].format(**data)
+
+    def get_execution_time(self):
+        """Get response time for request in milliseconds."""
+        start = flask.g.get('logconfig', {}).get('start')
+        execution_time = None
+
+        if start is not None:
+            # Only compute execution time once.
+            if 'execution_time' not in flask.g.logconfig:
+                flask.g.logconfig['execution_time'] = (
+                    milliseconds_between(start, datetime.datetime.now()))
+            execution_time = flask.g.logconfig['execution_time']
+
+        return execution_time
 
 
 def copy_current_request_context():
@@ -273,3 +304,10 @@ def request_context_from_record(record=None):
             yield ctx
     else:
         raise FlaskLogConfigException('No request context found on log record')
+
+
+def milliseconds_between(start, stop):
+    """Return milliseconds between `start` and `stop` datetime objects."""
+    diff = stop - start
+    return ((diff.days * 24 * 60 * 60 + diff.seconds) * 1000
+            + diff.microseconds / 1000.0)